+++ src/test/scala/base/UnitSpecBase.scala
@@ -1,0 +1,5 @@
+package base
+
+import org.scalatest.concurrent.Eventually
+
+trait UnitSpecBase extends WordSpecBase with Eventually
